3.6.0

+++ TMessagesProj/src/main/java/org/telegram/messenger/TopicsController.java
@@ -1005,7 +1005,7 @@ public class TopicsController extends BaseController {
         }
     }
 
-    public void loadTopic(long chatId, int topicId, Runnable runnable) {
+    public void loadTopic(long chatId, long topicId, Runnable runnable) {
         getMessagesStorage().loadTopics(-chatId, topics -> {
             AndroidUtilities.runOnUIThread(() -> {
                 if (BuildVars.LOGS_ENABLED) {
@@ -1019,7 +1019,7 @@ public class TopicsController extends BaseController {
                 } else {
                     ArrayList<TLRPC.TL_forumTopic> topicToReload = new ArrayList<>();
                     TLRPC.TL_forumTopic topic = new TLRPC.TL_forumTopic();
-                    topic.id = topicId;
+                    topic.id = (int) topicId;
                     reloadTopics(chatId, topicToReload, runnable);
                 }
             });
